Automatically install the latest release of OOD
When using these VMs for development/testing it is useful to be able to install the latest version of OOD. It would be nice to be able have a switch of some kind to automatically install the latest version of OOD. This commit adds a non-optional latest-ood install. Refining it could use host-environment variables to activate the fixed version / latest switch, or perhaps there are other ways to send args into a vagrant up command?